requirements.txt
asgiref==3.3.1
Django==3.1.6
django-autoslug==1.9.8
django-environ==0.4.5
djangorestframework==3.12.2
isort==5.7.0
pdfkit==0.6.1
Pillow==8.1.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
uritemplate==3.0.1
